+++ src/lib.rs
@@ -40,6 +40,12 @@
 //!
 //! # Crate features
 //!
+//! ## `std`
+//!
+//! *(on by default)*
+//!
+//! Enables support of types from `std`, such as the `Error` trait.
+//!
 //! ## `serde`
 //!
 //! *(off by default)*
@@ -74,6 +80,7 @@
 //! [`k256`]: https://docs.rs/k256/
 //! [docker-rng]: https://github.com/moby/moby/blob/master/pkg/namesgenerator/names-generator.go
 
+#![cfg_attr(not(feature = "std"), no_std)]
 // Documentation settings.
 #![doc(html_root_url = "https://docs.rs/elastic-elgamal/0.1.0")]
 // Linter settings.
@@ -93,11 +100,22 @@ mod proofs;
 mod serde;
 pub mod sharing;
 
+// Polyfill for `alloc` types.
+mod alloc {
+    #[cfg(not(feature = "std"))]
+    extern crate alloc;
+
+    #[cfg(not(feature = "std"))]
+    pub use alloc::{string::ToString, vec, vec::Vec};
+    #[cfg(feature = "std")]
+    pub use std::{string::ToString, vec, vec::Vec};
+}
+
 pub use crate::{
-    encryption::{Ciphertext, DiscreteLogTable, EncryptedChoice},
+    encryption::{ChoiceVerificationError, Ciphertext, DiscreteLogTable, EncryptedChoice},
     keys::{Keypair, PublicKey, PublicKeyConversionError, SecretKey},
     proofs::{
         LogEqualityProof, PreparedRange, ProofOfPossession, RangeDecomposition, RangeProof,
-        RingProof, RingProofBuilder,
+        RingProof, RingProofBuilder, VerificationError,
     },
 };