+++ src/proofs/ring.rs
@@ -6,13 +6,16 @@ use rand_core::{CryptoRng, RngCore};
 use serde::{Deserialize, Serialize};
 use subtle::ConstantTimeEq;
 
-use std::{fmt, mem};
+use core::{fmt, mem};
 
 #[cfg(feature = "serde")]
 use crate::serde::{ScalarHelper, VecHelper};
 use crate::{
-    encryption::ExtendedCiphertext, group::Group, proofs::TranscriptForGroup, Ciphertext,
-    PublicKey, SecretKey,
+    alloc::{vec, Vec},
+    encryption::ExtendedCiphertext,
+    group::Group,
+    proofs::{TranscriptForGroup, VerificationError},
+    Ciphertext, PublicKey, SecretKey,
 };
 
 /// An incomplete ring proving that the encrypted value is in the a priori known set of
@@ -111,10 +114,7 @@ impl<'a, G: Group> Ring<'a, G> {
             let dh_element = blinded_value - admissible_value;
             commitments = (
                 G::mul_generator(&response) - random_element * &challenge,
-                G::multi_mul(
-                    [response, -challenge].iter(),
-                    [log_base, dh_element].iter().copied(),
-                ),
+                G::multi_mul([&response, &-challenge], [log_base, dh_element]),
             );
         }
 
@@ -177,10 +177,7 @@ impl<'a, G: Group> Ring<'a, G> {
             let dh_element = self.ciphertext.blinded_element - admissible_value;
             let commitments = (
                 G::mul_generator(&response) - self.ciphertext.random_element * &challenge,
-                G::multi_mul(
-                    [response, -challenge].iter(),
-                    [log_base, dh_element].iter().copied(),
-                ),
+                G::multi_mul([&response, &-challenge], [log_base, dh_element]),
             );
 
             let mut eq_transcript = self.transcript.clone();
@@ -304,19 +301,20 @@ impl<G: Group> RingProof<G> {
         }
     }
 
-    #[must_use = "verification fail is returned as `false` and should be handled"]
     pub(crate) fn verify<'a>(
         &self,
         receiver: &PublicKey<G>,
         admissible_values: impl Iterator<Item = &'a [G::Element]> + Clone,
         ciphertexts: impl Iterator<Item = Ciphertext<G>>,
         transcript: &mut Transcript,
-    ) -> bool {
+    ) -> Result<(), VerificationError> {
         // Do quick preliminary checks.
         let total_rings_size: usize = admissible_values.clone().map(<[_]>::len).sum();
-        if total_rings_size != self.total_rings_size() {
-            return false;
-        }
+        VerificationError::check_lengths(
+            "items in all rings",
+            self.total_rings_size(),
+            total_rings_size,
+        )?;
 
         Self::initialize_transcript(transcript, receiver);
         // We add common commitments to the `transcript` as we cycle through rings,
@@ -349,8 +347,8 @@ impl<G: Group> RingProof<G> {
                         response,
                     ),
                     G::vartime_multi_mul(
-                        [response, &neg_challenge].iter().copied(),
-                        [receiver.element, dh_element].iter().copied(),
+                        [response, &neg_challenge],
+                        [receiver.element, dh_element],
                     ),
                 );
 
@@ -370,7 +368,11 @@ impl<G: Group> RingProof<G> {
         }
 
         let expected_challenge = transcript.challenge_scalar::<G>(b"c");
-        bool::from(expected_challenge.ct_eq(&self.common_challenge))
+        if expected_challenge.ct_eq(&self.common_challenge).into() {
+            Ok(())
+        } else {
+            Err(VerificationError::ChallengeMismatch)
+        }
     }
 
     pub(crate) fn total_rings_size(&self) -> usize {
@@ -500,7 +502,7 @@ mod tests {
     };
     use rand::{thread_rng, Rng};
 
-    use std::iter;
+    use core::iter;
 
     use super::*;
     use crate::group::{ElementOps, Ristretto};
@@ -537,15 +539,15 @@ mod tests {
             &mut transcript,
             &mut rng,
         );
-        let proof = RingProof::new(common_challenge, ring_responses);
 
-        let mut transcript = Transcript::new(b"test_ring_encryption");
-        assert!(proof.verify(
-            keypair.public(),
-            iter::once(&admissible_values as &[_]),
-            iter::once(ciphertext),
-            &mut transcript
-        ));
+        RingProof::new(common_challenge, ring_responses)
+            .verify(
+                keypair.public(),
+                iter::once(&admissible_values as &[_]),
+                iter::once(ciphertext),
+                &mut Transcript::new(b"test_ring_encryption"),
+            )
+            .unwrap();
 
         // Check a proof for encryption of 1.
         let value = Ristretto::generator();
@@ -571,15 +573,15 @@ mod tests {
             &mut transcript,
             &mut rng,
         );
-        let proof = RingProof::new(common_challenge, ring_responses);
 
-        let mut transcript = Transcript::new(b"test_ring_encryption");
-        assert!(proof.verify(
-            keypair.public(),
-            iter::once(&admissible_values as &[_]),
-            iter::once(ciphertext),
-            &mut transcript
-        ));
+        RingProof::new(common_challenge, ring_responses)
+            .verify(
+                keypair.public(),
+                iter::once(&admissible_values as &[_]),
+                iter::once(ciphertext),
+                &mut Transcript::new(b"test_ring_encryption"),
+            )
+            .unwrap();
     }
 
     #[test]
@@ -616,15 +618,15 @@ mod tests {
                 &mut transcript,
                 &mut rng,
             );
-            let proof = RingProof::new(common_challenge, ring_responses);
 
-            let mut transcript = Transcript::new(b"test_ring_encryption");
-            assert!(proof.verify(
-                keypair.public(),
-                iter::once(admissible_values.as_slice()),
-                iter::once(ciphertext),
-                &mut transcript
-            ));
+            RingProof::new(common_challenge, ring_responses)
+                .verify(
+                    keypair.public(),
+                    iter::once(admissible_values.as_slice()),
+                    iter::once(ciphertext),
+                    &mut Transcript::new(b"test_ring_encryption"),
+                )
+                .unwrap();
         }
     }
 
@@ -669,15 +671,15 @@ mod tests {
 
             let common_challenge =
                 Ring::aggregate(rings, keypair.public().element, &mut transcript, &mut rng);
-            let proof = RingProof::new(common_challenge, ring_responses);
 
-            let mut transcript = Transcript::new(b"test_ring_encryption");
-            assert!(proof.verify(
-                keypair.public(),
-                iter::repeat(&admissible_values as &[_]).take(RING_COUNT),
-                ciphertexts.into_iter(),
-                &mut transcript,
-            ));
+            RingProof::new(common_challenge, ring_responses)
+                .verify(
+                    keypair.public(),
+                    iter::repeat(&admissible_values as &[_]).take(RING_COUNT),
+                    ciphertexts.into_iter(),
+                    &mut Transcript::new(b"test_ring_encryption"),
+                )
+                .unwrap();
         }
     }
 
@@ -741,16 +743,16 @@ mod tests {
 
             let common_challenge =
                 Ring::aggregate(rings, keypair.public().element, &mut transcript, &mut rng);
-            let proof = RingProof::new(common_challenge, ring_responses);
             let admissible_values = admissible_values.iter().map(|values| values as &[_]);
 
-            let mut transcript = Transcript::new(b"test_ring_encryption");
-            assert!(proof.verify(
-                keypair.public(),
-                admissible_values,
-                ciphertexts.into_iter(),
-                &mut transcript
-            ));
+            RingProof::new(common_challenge, ring_responses)
+                .verify(
+                    keypair.public(),
+                    admissible_values,
+                    ciphertexts.into_iter(),
+                    &mut Transcript::new(b"test_ring_encryption"),
+                )
+                .unwrap();
         }
     }
 
@@ -774,13 +776,14 @@ mod tests {
         let ciphertexts: Vec<_> = (0..5)
             .map(|i| builder.add_value(&admissible_values, i & 1).inner)
             .collect();
-        let proof = RingProof::new(builder.build(), ring_responses);
 
-        assert!(proof.verify(
-            keypair.public(),
-            iter::repeat(&admissible_values as &[_]).take(5),
-            ciphertexts.into_iter(),
-            &mut Transcript::new(b"test_ring_encryption"),
-        ));
+        RingProof::new(builder.build(), ring_responses)
+            .verify(
+                keypair.public(),
+                iter::repeat(&admissible_values as &[_]).take(5),
+                ciphertexts.into_iter(),
+                &mut Transcript::new(b"test_ring_encryption"),
+            )
+            .unwrap();
     }
 }