+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 - Update `elliptic-curve` dependency.
 
+- Return new error types (`VerificationError` and `ChoiceVerificationError`) on proof
+  verification failure instead of signalling failure via `bool` or `Option` values.
+
 ### Fixed
 
 - Remove unused `byteorder` and `smallvec` dependencies.